cmd/cluster-apply: Rephrase description
"Apply configuration changes" doesn't accurately describe what the
command does. It should be clear that the command deploys a cluster
when no cluster is present or updates a cluster when one is.

"with components" is confusing. It makes it sound like the command
only works on clusters with components, which is not what we're
trying to convey. The long description explains this behavior.